Keep expanded state in watch widget
