+++ scrutiny/sdk/docs/source/_static/code-examples/hil_testing/hil_testing_1_powerup_check.py
@@ -1,5 +1,3 @@
-import scrutiny
-import scrutiny.sdk
 from scrutiny.sdk.client import ScrutinyClient
 
 hostname = 'localhost'
